backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-dotenv==1.0.1
python-multipart==0.0.12
openai==1.54.3
pypdf==5.1.0
python-pptx==1.0.2
Pillow==10.4.0
httpx==0.27.2
reportlab==4.2.5
